+++ commons/src/androidMain/kotlin/test/TestBalloonCommons.android.kt
@@ -1,3 +1,3 @@
 package at.asitplus.testballoon
 
-actual  val defaultMaxLen : Int = 10
+internal actual val defaultMaxLen : Int = 10