+++ internals/src/androidMain/kotlin/test/TestBalloonInternals.android.kt
@@ -1,0 +1,4 @@
+package at.asitplus.testballoon
+
+
+actual var totalMaxLen: Int = 200