+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Prefixes for data test and property test
 * Shorter compact names
 * Global test name length configuration -> deprecate `DEFAULT_TEST_NAME_MAX_LEN`
+* Hard fails on too long overall test names on Android
 
 ### 0.6.2-RC
 * Support TestBalloon 0.8-RC